Made changes in the code as per feedback on PR
