+++ portfolio-project-react/src/App.js
@@ -1,7 +1,7 @@
 import "./index.css";
-import Header from "./components/Header";
-import Content from "./components/Content";
-import Footer from "./components/Footer";
+import Header from "./components/Header/Header";
+import Content from "./components/Content/Content";
+import Footer from "./components/Footer/Footer";
 
 const App = () => {
   return (
