--- portfolio-project-react/src/components/Content.jsx
@@ -1,15 +1,0 @@
-import Details from './Details';
-import Picture from './Picture';
-import logo from '../assets/ic-profile.png'
-
-
-const Content = () => {
-  return (
-    <main>
-      <Details />
-      <Picture url={logo} size={'250px'}/>
-    </main>
-  );
-};
-
-export default Content;
