+++ portfolio-project-react/src/components/Content/Content.module.css
@@ -1,0 +1,6 @@
+main {
+  display: flex;
+  margin: 8%;
+  box-sizing: border-box;
+  gap: 10%;
+}
