--- portfolio-project-react/src/components/Details.jsx
@@ -1,31 +1,0 @@
-const Details = () => {
-  return (
-    <div className='details-section'>
-      <h1 class='title-about'>About</h1>
-      <p id='intro-data'>
-        I am a highly experienced Full Stack Web Developer with a proven track
-        record of developing, designing, and integrating robust Web Applications
-        using the latest technologies like React JS, Express JS, Node, MongoDB,
-        etc. I have in-depth knowledge in this area and can provide responsive
-        web application development services. I am also familiar with common
-        framework APIs, dedicated to creating user-centric and innovative
-        web-based solutions to drive business growth.
-      </p>
-
-      <hr class='seperator' />
-
-      <h3 class='inline-div'>Contact:</h3>
-      <p class='inline-div'>+92 333 1234567</p>
-      <br />
-      <h3 class='inline-div'>Email:</h3>
-      <p class='inline-div'>sana.iqbal@gmail.com</p>
-      <div id='details'>
-        <div>
-          <h1>Work Experience</h1>
-        </div>
-      </div>
-    </div>
-  );
-};
-
-export default Details;
