+++ portfolio-project-react/src/components/Details/Details.module.css
@@ -1,0 +1,15 @@
+.detailsSection {
+  width: 70%;
+}
+
+.titleAbout {
+  margin-top: 0;
+}
+
+.inlineDiv {
+  display: inline-block;
+  padding-right: 10px;
+}
+.details {
+  margin-top: 50px;
+}
