+++ portfolio-project-react/src/components/Details/constants.jsx
@@ -1,0 +1,3 @@
+export const ABOUT = "I am a highly experienced Full Stack Web Developer with a proven track record of developing, designing, and integrating robust Web Applications using the latest technologies like React JS, Express JS, Node, MongoDB, etc. I have in-depth knowledge in this area and can provide responsive web application development services. I am also familiar with common framework APIs, dedicated to creating user-centric and innovative web-based solutions to drive business growth.";
+export const CONTACT = "+92 333 1234567";
+export const EMAIL = "sana.iqbal@gmail.com"
