+++ portfolio-project-react/src/components/Footer/Footer.jsx
@@ -1,3 +1,4 @@
+import styles from "./Footer.module.css"
 const Footer = () => {
   return (
     <footer>
