+++ portfolio-project-react/src/components/Footer/Footer.module.css
@@ -1,0 +1,20 @@
+footer {
+  width: 100%;
+  background-color: rgb(1, 1, 65);
+  color: white;
+  padding: 20px 30px 20px 30px;
+  display: flex;
+  align-items: center;
+  bottom: 0px;
+  box-sizing: border-box;
+}
+
+footer a {
+  cursor: pointer;
+  font-weight: 600;
+  font-size: x-large;
+}
+
+footer a:hover {
+  color: rgb(195, 195, 255);
+}
