+++ portfolio-project-react/src/components/Header/Header.jsx
@@ -1,3 +1,5 @@
+import styles from "./Header.module.css"
+
 const Header = () => {
   return (
     <header>
