+++ portfolio-project-react/src/components/Picture/Picture.module.css
@@ -1,0 +1,27 @@
+.imageContainer {
+  overflow: hidden;
+}
+
+.pictureSection {
+  width: 30%;
+  height: auto;
+  background-color: rgb(213, 213, 236);
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  padding: 3%;
+  box-sizing: border-box;
+}
+
+.pictureSection img {
+  max-width: 250px;
+  max-height: 250px;
+}
+
+#name {
+  margin-top: 50px;
+}
+
+#jobTitle {
+  margin-top: 0px;
+}
