+++ portfolio-project-react/src/components/Picture/constants.jsx
@@ -1,0 +1,2 @@
+export const NAME = "Sana Iqbal";
+export const JOBTITLE = "Software Engineer";
